Upgrade webpack to version 4 (#74)
* refactor: Upgrade Webpack to version 4

The current version of Webpack is 4 and we are using 3. This commit
upgrades our configurations to version 4. Note some dependencies are
still in alpha stage, but by the time we release sheikah, for sure those
depencencies will be already stable.

* style: move dev dependencies into its section

move commitlint and electron debug dependencies to devDependencies section of package.json

* style: add mode "production" to webpack's base config

by default webpack uses the "production" mode if no mode is specified
but emits a warning to the console. I'm setting the mode explicitly to
remove such warning.