feat(ipc-renderer): stateless ipc client
add an ipc client to be used to interact with handlers defined in
Electron's Main process.

this client follows a stateless design, that is, on sending a request
the client will bound once to a special and unique "reply channel"
that will be used by the Main process to univocally send the response
for that request. if there's any problem parsing or decoding the
request, the response will be sent to an special "dead letter"
channel, is responsability of the frontend (not the client) to
subscribe to that channel in order to be notified of invalid requests
that are not being processed correctly.

the client also supports "notifications", that is, requests that do
not wait for a response.

example:

```typescript
import { Client } from "app/renderer/api"

const client = new Client()

client.request("echo", "this", "are", "parameters")
  .then(processResult)
  .catch(processErrors)

client.notify("echo", "this", "are", "parameters")
  .then(processResult)
  .catch(processErrors)
```