feat(actions): modify return types of unlockWallet action creator
UnlockWallet action creator returns the wallet to the container in the case of a successful unlock.
It returns an error message in the case of an unsuccessful unlock. The error message will be
specific in the case of a handler error and generic in the case of any IPC related error.

re #282