feat(actions): remove unlockWallet "action creator" function
This was not a pure action creator as thunk understood it as it did not return or dispatch any
action. It was just used to get the services from the thunk middleware. This was not very clear and
this obscure logic has been simplified by getting services as props in the container.

re #311