feat(wallet): final key now contains outpoints to utxos and stxos
Actual outputs and inputs are not in the wallet runtime data structure, but instead they will be in
stored in the transaction data structure (which is yet to be defined). The final keys in the wallet
will store outpoints to utxos and stxos, that is, pointers to transaction outputs.

re #398